requirements.txt
Django==4.2.18
django-environ==0.11.2
psycopg2-binary==2.9.9
whitenoise==6.7.0
Pillow==10.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-allauth==0.63.6
django-browser-reload==1.12.1
gunicorn==21.2.0
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
